requirements.txt
biopython
bcbio-gff